requirements.txt
GitPython==3.1.31
python-dotenv~=1.0.0